backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pymysql==1.0.3
alembic==1.10.4
python-dotenv==1.0.0
email-validator==2.0.0
pillow==9.5.0
pytest==7.3.1 